tweeter/requirements.txt
psycopg2-binary==2.9.7
twikit==2.3.3
XClientTransaction==1.0.2
beautifulsoup4
matplotlib==3.10.8
seaborn==0.13.2
fonttools[woff]
brotli==1.2.0
Pillow==12.2.0
